Better determination of which columns' resize handle to enable
Even if a column is not resizable, if other columns are resizable, then
the resize handle can still be used to push this column around while
resizing other columns.

Also, the last resizable column should not have a resize handle if
forceFitColumns. This rule applies for the right-most resizable column,
not just the right-most column, since even if a column is not the
right-most column, if all the columns on the right of it are not
resizable, then this column's resize handle can't do anything.

Finally, these resize rules should persist even after columns are
reordered and column visibility is changed.