Secondary header row support (usually used for filters).
The DOM element will always be created; the display and height are controlled by options.showSecondaryHeaderRow and options.secondaryHeaderRowHeight.  The content of the element is to be populated externally.  The new onBeforeDestroy event is supplied so that the user can do cleanup for the content of the element.

New properties:
- showSecondaryHeaderRow
- secondaryHeaderRowHeight

New methods:
- getSecondaryHeaderRow
- showSecondaryHeaderRow
- hideSecondaryHeaderRow

New events:
- onBeforeDestroy