Use local variables instead of the data object to store current getItem and getLength function references.
This is an issue if a parent window passes an array to a child window
(e.g. a popup). Slick grid in the child window will see that the data
object does not have getItem and getLength functions so it will add them.
However, when the child window closes, parent window will have an array
that has "getItem" and "getLength" properties pointing to functions
that were defined in a now-nonexistent window. If it now passes the same
data object to another child window, things will go BOOOM.

Using local variables prevents this problem because the data object
is not being changed in child window.