This checkin adds cell highlighting functionality.  To set highlighted cells, call grid.setHighlightedCells and pass in an object that will be queries to determine whether a cell at a given row and column id is highlighted (cellsToHighlight[row][columnId]).  Subsequent calls to setHighlightedCells will reset highlighting on all cells previously highlighted.  A CSS class to be applied to the cells can be set in options.cellHighlightCssClass.
Highlighting is really fast since it only has to touch the nodes that are currently in the DOM tree, and that is limited by the viewport.  Setting highlighting for 50'000 rows is just as fast as for 50.

This feature can be used to provide change tracking hints or highlight changed data as it is being pulled from the server in real time (poll or push).

- NEW:  added "setHighlightedCells" grid method.
- NEW:  added "cellHighlightCssClass" grid option.