DataView grouping perf
Optimized DataView grouping by dynamically compiling the aggregator
accumulation loop and inlining the accumulator function.  This removes
the function call overhead and considerably speeds up aggregate
calculation for large numbers of items.  One caveat is that aggregators
now have to be self-contained and not reference any variables outside
of the "this" scope.