remove two step css rule creation
This has stopped working in google chrome. Moreover, I do not see why it
was done like this in the first place as the rules can be gotten right
from the style object after it is populated with text.

TODO long term this is too fragile and should be done a better way.