Add failing test case for inlined filtering.
Filters with `return true|false` statements that are not terminated by
semi-colons or closing braces will be left dangling in the
slickgrid-compiled function. This causes invalid data to be returned when
filtering.

This can be an issue when JS compression/minification is integrated into a
project, because some minifiers strip trailing semicolons when possible.