Added call to "handleScroll" on `updateRowCount()`
since a disappearing vetical scrollbar can cause problems in internet explorer. When the vertical scollbar disappears Chrome
and firefox will fire a scroll event for the horizontal direction (because now there is more space in the div) but the
internet explorer does not. The call to handleScroll fixes this problem.