Make aggregators "accumulate" function optional
Accumulators may not need/want to iterate through all the rows, and thus not use the "accumulate" function. This change checks for the presence of that function, and if it does not exist, then it does not do the iteration. A "CountAggregator" has also been added as a feature, but also to show how/why this is necessary. Other use cases are "count distinct" or perhaps aggregations are already done and stored elsewhere for quick use.