Bug in getRowDiffs() causing incorrect filter
In getRowDiffs(), the "to" variable is set to newRows.length. As an example, if newRows.length is 1 and rows.length is 2, the for loop only iterates from 0 to 1, and thus not detecting any difference so long as the first item in both newRows and rows remain the same.

I also had to add a check for !item. In the example above, newRows[1] would not exist.