Fix error for $footerRowTarget undefined
If there is no footer row on the grid, and you call grid.getFooterRowColumn(colId), then you'll get an error message. I realize that you shouldn't be calling this function if you don't have a footer, but I did accidentally and uncovered this simple fix in case others do the same.